refactor: Change DomainException package and update imports
