+++ domain/model/src/main/java/co/com/crediya/model/user/exceptions/LastNameRequiredException.java
@@ -1,6 +1,6 @@
 package co.com.crediya.model.user.exceptions;
 
-import co.com.crediya.model.common.exceptions.DomainException;
+import co.com.crediya.model.common.exception.DomainException;
 
 public class LastNameRequiredException extends DomainException {
     public LastNameRequiredException() {
