+++ domain/usecase/src/main/java/co/com/crediya/exception/NotFoundException.java
@@ -1,6 +1,6 @@
 package co.com.crediya.exception;
 
-import co.com.crediya.model.common.exceptions.DomainException;
+import co.com.crediya.model.common.exception.DomainException;
 
 public abstract class NotFoundException extends DomainException {
     protected NotFoundException(String code, String message) {
